Add note about disabling pre-existing last-edit
One of the Vim installations that I use already had last-edit functionality, and it triggered on Git commit messages, so I thought that something was wrong with vim-lastplace. It took me a while to track down where it was coming from. This paragraph may save others time if they're in a similar situation. If you'd rather not include it (because it's too much detail or is out of scope for vim-lastplace's documentation), then that's fine too.